[Feat] 게스트하우스 세부정보 API 제작

+++ backend/src/main/java/com/backend/repository/GuesthouseRepository.java
@@ -4,5 +4,5 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import com.backend.entity.Guesthouse;
 
-public interface GuesthouseRepository extends JpaRepository<Guesthouse, Long> {
+public interface GuesthouseRepository extends JpaRepository<Guesthouse, Integer> {
 }
